+++ code/renderercommon/tr_image_bmp.c
@@ -54,7 +54,7 @@ void R_LoadBMP( const char *name, byte **pic, int *width, int *height )
 		byte *b;
 		void *v;
 	} buffer;
-	int		length;
+	unsigned	length;
 	BMPHeader_t bmpHeader;
 	byte		*bmpRGBA;
 
@@ -164,7 +164,7 @@ void R_LoadBMP( const char *name, byte **pic, int *width, int *height )
 	numPixels = columns * rows;
 
 	if(columns <= 0 || !rows || numPixels > 0x1FFFFFFF // 4*1FFFFFFF == 0x7FFFFFFC < 0x7FFFFFFF
-	    || ((numPixels * 4) / columns) / 4 != rows)
+	    || (((int)numPixels * 4) / columns) / 4 != rows)
 	{
 	  ri.Error (ERR_DROP, "LoadBMP: %s has an invalid image size", name);
 	}
